requirements.txt
Django~=3.1.4
PyGithub~=1.54
python-dotenv~=0.15.0
requests~=2.25.1
pyrebase5~=5.0.1
gunicorn~=20.0.4
pyttanko~=1.2.1
bs4~=0.0.1
beautifulsoup4~=4.9.3